refactor: separate Layers panel presentation from lifecycle actions
